pyproject.toml
[tool.poetry]
name = "meeting-transcription"
version = "1.0.0"
description = "Meeting transcription service with AI-powered summarization"
authors = ["Kurt Niemi <kurtn@lll-solutions.com>"]
readme = "README.md"
packages = [{include = "meeting_transcription", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"

# Web Framework
flask = "^3.0.0"

# HTTP Client
requests = "^2.31.0"
svix = "^1.0.0"                      # For Recall.ai webhook signature verification

# Environment Variables
python-dotenv = "^1.0.0"

# LLM Providers
google-genai = "^1.56.0"                 # For Google Vertex AI (Gemini)
aisuite = {git = "https://github.com/kurtn718/aisuite.git", rev = "add-gemini3-support", extras = ["openai", "anthropic", "google"]}

# PDF Generation
weasyprint = "^68.0"                 # HTML/CSS to PDF
markdown = "^3.5.0"                  # Markdown processing

# Google Cloud (for GCP deployment)
google-cloud-storage = "^2.14.0"
google-cloud-firestore = "^2.14.0"
google-cloud-secret-manager = "^2.18.0"
google-cloud-tasks = "^2.14.0"

# Firebase Auth (for user authentication)
firebase-admin = "^6.2.0"

# Utilities
schedule = "^1.2.0"                  # For scheduled tasks
gunicorn = "^22.0.0"                 # Production WSGI server
bcrypt = "^5.0.0"

# Security
flask-limiter = "^3.5.0"            # Rate limiting
redis = "^5.0.0"                     # Redis client for distributed rate limiting (optional)

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
ruff = "^0.1.0"
mypy = "^1.7.0"
types-requests = "^2.31.0"
types-markdown = "^3.5.0"
pip-audit = "^2.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 100

select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort (import sorting)
    "N",      # pep8-naming
    "UP",     # pyupgrade (modern Python syntax)
    "B",      # flake8-bugbear (common bugs)
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "RUF",    # Ruff-specific rules
]

ignore = [
    "E501",   # Line too long (let formatter handle it)
]

exclude = [
    ".git",
    "__pycache__",
    "venv",
    ".venv",
    "*.egg-info",
]

[tool.ruff.per-file-ignores]
"tests/**/*.py" = [
    "S101",   # Allow assert in tests
]
"main.py" = ["E402"]
"src/meeting_transcription/pipeline/chunkers/educational_chunker.py" = ["E402"]
"src/meeting_transcription/pipeline/factories/pipeline_factory.py" = ["E402"]
"src/meeting_transcription/pipeline/formatters/study_guide_formatter.py" = ["E402"]
"src/meeting_transcription/pipeline/prompts/educational_prompts_engine.py" = ["E402"]

[tool.mypy]
python_version = "3.12"
mypy_path = "src"
explicit_package_bases = true
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false
warn_no_return = true
check_untyped_defs = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "anthropic.*",
    "aisuite.*",
    "google.cloud.*",
    "google.auth.*",
    "google.oauth2.*",
    "weasyprint.*",
    "schedule.*",
    "firebase_admin.*",
    "svix.*",
    "flask_limiter.*",
    "flask_cors.*",
    "redis.*",
    "bcrypt.*",
    "pipeline.*",
    "educational_prompts",
    "src.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "meeting_transcription.api.*",
    "meeting_transcription.services.*",
    "meeting_transcription.providers.*",
    "meeting_transcription.plugins.*",
    "meeting_transcription.utils.*",
    "meeting_transcription.pipeline.*",
    "meeting_transcription.google_meet.*",
    "meeting_transcription.models.*",
]
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=src",
    "--cov=main",
    "--cov-report=term-missing",
    "--cov-report=html",
]
